refactor: update codebase
